+++ src/core/util/array_tim_sorter.rs
@@ -60,9 +60,8 @@ impl<T, C: Comparator<T>> Sorter for ArrayTimSorter<'_, T, C>
 where
     T: Copy,
 {
-    fn compare(&mut self, i: i32, j: i32) -> Result<i32> {
-        self.comparator
-            .compare(&self.arr[i as usize], &self.arr[j as usize])
+    fn compare(&mut self, i: usize, j: usize) -> Result<i32> {
+        self.comparator.compare(&self.arr[i], &self.arr[j])
     }
 
     fn swap(&mut self, i: usize, j: usize) -> Result<()> {
@@ -76,7 +75,7 @@ where
     }
 
     fn compare_pivot(&mut self, j: i32) -> Result<i32> {
-        self.compare(self.pivot_index, j)
+        self.compare(self.pivot_index as usize, j as usize)
     }
 }
 impl<T, C: Comparator<T>> TimSorterBase for ArrayTimSorter<'_, T, C>
